@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.19.2",
-    "bootstrap": "^4.6.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.0.1",
-    "redux": "^4.1.0",
+    "axios": "^0.23.0",
+    "bootstrap": "^5.1.3",
+    "fsevents": "^2.3.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.5",
+    "react-router-dom": "^5.3.0",
+    "react-scripts": "^4.0.3",
+    "redux": "^4.1.1",
     "redux-thunk": "^2.3.0"
   },
   "scripts": {
@@ -36,6 +37,6 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0"
+    "@testing-library/react": "^12.1.2"
   }
 }
