@@ -1,8 +1,12 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import "./nav.css";
+import { useSelector, useDispatch } from "react-redux";
+import { logoutUser } from '../../redux/actions/authActions'
 
 const Nav = (props) => {
+    const auth = useSelector((state) => state.auth.email)
+    const dispatch = useDispatch()
     return (
         <div className="nav-strip">
             <Link to={"/order"} className="nav-link">
@@ -15,9 +19,10 @@ const Nav = (props) => {
                     <label className="nav-label">View Orders</label>
                 </div>
             </Link>
-            <Link to={"/login"} className="nav-link">
+            <Link to={"/"} className="nav-link">
                 <div className="nav-link-style">
-                    <label className="nav-label">Log Out</label>
+                    {auth} | {" "}
+                    <label onClick={() => dispatch(logoutUser())} className="nav-label">Log Out</label>
                 </div>
             </Link>
         </div>
