@@ -1,4 +1,4 @@
-const windowsPlatforms = ['Win32', 'Win64', 'Windows', 'WinCE'];
+// const windowsPlatforms = ['Win32', 'Win64', 'Windows', 'WinCE'];
 
 export const SERVER_IP = "http://localhost:4000";
 
